chore(clerk-js): Organize all tests in __tests__ directories
