+++ packages/clerk-js/src/ui/foundations/__tests__/createInternalTheme.test.ts
@@ -1,4 +1,4 @@
-import { createInternalTheme } from './createInternalTheme';
+import { createInternalTheme } from '../createInternalTheme';
 
 describe('createInternalTheme', () => {
   it('handles empty objects', () => {
