+++ packages/clerk-js/src/ui/utils/__tests__/phoneUtils.test.ts
@@ -4,7 +4,7 @@ import {
   getCountryIsoFromFormattedNumber,
   getFlagEmojiFromCountryIso,
   getLongestValidCountryCode,
-} from './phoneUtils';
+} from '../phoneUtils';
 
 describe('phoneUtils', () => {
   describe('countryIsoToFlagEmoji(iso)', () => {
