+++ packages/clerk-js/src/utils/__tests__/jwt.test.ts
@@ -1,6 +1,6 @@
 import jwtGen from 'jsonwebtoken';
 
-import { decode } from './jwt';
+import { decode } from '../jwt';
 
 const jwt = jwtGen.sign(
   {
