+++ packages/clerk-js/src/utils/__tests__/path.test.ts
@@ -1,4 +1,4 @@
-import { joinPaths } from './path';
+import { joinPaths } from '../path';
 
 describe('joinPaths(a, b)', () => {
   const cases = [
