+++ packages/clerk-js/src/utils/__tests__/resourceParams.test.ts
@@ -1,6 +1,6 @@
 import { UpdateUserParams } from '@clerk/types';
 
-import { normalizeUnsafeMetadata } from './resourceParams';
+import { normalizeUnsafeMetadata } from '../resourceParams';
 
 describe('normalizeUnsafeMetadata', () => {
   it('handles unsafeMetadata', () => {
