+++ packages/clerk-js/src/utils/__tests__/url.test.ts
@@ -11,7 +11,7 @@ import {
   removeSearchParameterFromHash,
   setSearchParameterInHash,
   trimTrailingSlash,
-} from './url';
+} from '../url';
 
 describe('isAccountsHostedPages(url)', () => {
   const goodUrls: Array<[string | URL, boolean]> = [
